+++ noodles-csi/src/binning_index.rs
@@ -4,10 +4,10 @@ mod reference_sequence_ext;
 
 pub use reference_sequence_ext::ReferenceSequenceExt;
 
-use std::{io, ops::RangeBounds};
+use std::io;
 
 use noodles_bgzf as bgzf;
-use noodles_core::Position;
+use noodles_core::region::Interval;
 
 use super::index::reference_sequence::bin::Chunk;
 
@@ -23,9 +23,9 @@ pub trait BinningIndex {
     fn unplaced_unmapped_record_count(&self) -> Option<u64>;
 
     /// Returns the chunks that overlap with the given region.
-    fn query<B>(&self, reference_sequence_id: usize, interval: B) -> io::Result<Vec<Chunk>>
+    fn query<I>(&self, reference_sequence_id: usize, interval: I) -> io::Result<Vec<Chunk>>
     where
-        B: RangeBounds<Position> + Clone;
+        I: Into<Interval>;
 
     /// Returns the start position of the first record in the last linear bin.
     ///